Simplify manifest
