+++ src/collection.hpp
@@ -51,7 +51,7 @@ public:
 
     virtual void match(std::vector<event> &events /* output */,
         std::unordered_set<std::string_view> &seen_actions /* input & output */,
-        const object_store &store, const ddwaf::manifest &manifest, collection_cache &cache,
+        const object_store &store, collection_cache &cache,
         const std::unordered_set<rule::ptr> &rules_to_exclude,
         const std::unordered_map<rule::ptr, object_set> &objects_to_exclude,
         ddwaf::timer &deadline) const;
@@ -79,7 +79,7 @@ public:
 
     void match(std::vector<event> &events /* output */,
         std::unordered_set<std::string_view> &seen_actions /* input & output */,
-        const object_store &store, const ddwaf::manifest &manifest, collection_cache &cache,
+        const object_store &store, collection_cache &cache,
         const std::unordered_set<rule::ptr> &rules_to_exclude,
         const std::unordered_map<rule::ptr, object_set> &objects_to_exclude,
         ddwaf::timer &deadline) const override;
