+++ src/exclusion/rule_filter.cpp
@@ -17,15 +17,10 @@ rule_filter::rule_filter(
     for (auto it = rule_targets.begin(); it != rule_targets.end();) {
         rule_targets_.emplace(std::move(rule_targets.extract(it++).value()));
     }
-
-    for (auto &cond : conditions) {
-        const auto &cond_targets = cond->get_targets();
-        targets_.insert(cond_targets.begin(), cond_targets.end());
-    }
 }
 
 std::unordered_set<rule::ptr> rule_filter::match(const object_store &store,
-    const ddwaf::manifest &manifest, cache_type &cache, ddwaf::timer &deadline) const
+    cache_type &cache, ddwaf::timer &deadline) const
 {
     if (cache.result) {
         return {};
@@ -47,7 +42,7 @@ std::unordered_set<rule::ptr> rule_filter::match(const object_store &store,
         }
 
         // TODO: Condition interface without events
-        auto opt_match = cond->match(store, manifest, {}, run_on_new, deadline);
+        auto opt_match = cond->match(store, {}, run_on_new, deadline);
         if (!opt_match.has_value()) {
             cached_result->second = false;
             return {};
