+++ src/ruleset.hpp
@@ -32,49 +32,6 @@ struct ruleset {
         } else {
             priority_collections[rule->get_tag("type")].insert(rule);
         }
-        rules_by_type[rule->get_tag("type")].emplace(rule);
-        rules_by_category[rule->get_tag("category")].emplace(rule);
-    }
-
-    std::set<rule::ptr> get_rules_by_type(std::string_view type) const
-    {
-        auto it = rules_by_type.find(type);
-        if (it == rules_by_type.end()) {
-            return {};
-        }
-        return it->second;
-    }
-
-    std::set<rule::ptr> get_rules_by_category(std::string_view category) const
-    {
-        auto it = rules_by_category.find(category);
-        if (it == rules_by_category.end()) {
-            return {};
-        }
-        return it->second;
-    }
-
-    std::set<rule::ptr> get_rules_by_type_and_category(
-        std::string_view type, std::string_view category) const
-    {
-        auto type_it = rules_by_type.find(type);
-        if (type_it == rules_by_type.end()) {
-            return {};
-        }
-
-        auto category_it = rules_by_category.find(category);
-        if (category_it == rules_by_category.end()) {
-            return {};
-        }
-
-        const std::set<rule::ptr> &type_set = type_it->second;
-        const std::set<rule::ptr> &category_set = category_it->second;
-        std::set<rule::ptr> intersection;
-
-        std::set_intersection(type_set.begin(), type_set.end(), category_set.begin(),
-            category_set.end(), std::inserter(intersection, intersection.begin()));
-
-        return intersection;
     }
 
     ddwaf_object_free_fn free_fn{ddwaf_object_free};
@@ -92,9 +49,6 @@ struct ruleset {
     std::unordered_map<std::string_view, collection> collections;
 
     ddwaf::rule_data::dispatcher dispatcher;
-
-    std::unordered_map<std::string_view, std::set<rule::ptr>> rules_by_type;
-    std::unordered_map<std::string_view, std::set<rule::ptr>> rules_by_category;
 };
 
 } // namespace ddwaf
